+++ packages/core-js/actual/json/parse.js
@@ -1,6 +1,5 @@
 'use strict';
-require('../../modules/es.object.keys');
+var parent = require('../../stable/json/parse');
 require('../../modules/esnext.json.parse');
-var path = require('../../internals/path');
 
-module.exports = path.JSON.parse;
+module.exports = parent;